refactor(me): split /me/page.tsx into _components
Extract ProfileHero, StatsGrid, Stat, Heatmap, and MyCoursesSection so
the page itself stays focused on data orchestration.
